- using Feature2D everywhere to unify the interface ...
	modified:   ../cvt/CMakeLists.txt
	modified:   ../cvt/vision/AGAST.cpp
	modified:   ../cvt/vision/AGAST.h
	modified:   ../cvt/vision/Brief.cpp
	modified:   ../cvt/vision/Brief.h
	modified:   ../cvt/vision/FAST.cpp
	modified:   ../cvt/vision/FAST.h
	deleted:    ../cvt/vision/FeatureExtractor.h
	modified:   ../cvt/vision/Ferns.cpp
	modified:   ../cvt/vision/Ferns.h
	modified:   ../cvt/vision/PatchGenerator.cpp
	modified:   ../cvt/vision/PatchGenerator.h
	modified:   ../cvt/vision/agast/agast5_8.cpp
	modified:   ../cvt/vision/agast/agast7_12d.cpp
	modified:   ../cvt/vision/agast/agast7_12s.cpp
	modified:   ../cvt/vision/agast/oast9_16.cpp
	modified:   ../cvt/vision/fast/fast10.cpp
	modified:   ../cvt/vision/fast/fast11.cpp
	modified:   ../cvt/vision/fast/fast12.cpp
	modified:   ../cvt/vision/fast/fast9.cpp
	modified:   ../cvt/vision/fast/fast9simd.cpp
	modified:   ../tests/brief/brief.cpp
	modified:   ../tests/features/FeatureTracker.cpp
	modified:   ../tests/features/FeatureTracker.h
	modified:   ../tests/features/FeatureView.h
	modified:   ../tests/features/features.cpp
	modified:   ../tests/ferns/ferns.cpp

+++ cvt/vision/PatchGenerator.cpp
@@ -43,7 +43,7 @@ namespace cvt
 		_affine = R_theta * R_phi.transpose() * S * R_phi;
 	}
 	
-	void PatchGenerator::next( Image & outputPatch, const cvt::Image & inputImage, const Eigen::Vector2i & patchCenter )
+	void PatchGenerator::next( Image & outputPatch, const cvt::Image & inputImage, const Vector2f & patchCenter )
 	{
 		// generate new random affine Transform
 		randomizeAffine();
